new classes for table and listBox
