+++ aka/task3/src/main/java/mentee/DataTableBean.java
@@ -1,0 +1,15 @@
+package mentee;
+
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * Created by TheFallen on 16.12.15.
+ */
+public class DataTableBean extends ListBean {
+
+    public List<String> getListDataValue() {
+        return listDataValue;
+    }
+}
